+++ perception/lidar_centerpoint/include/lidar_centerpoint/centerpoint_trt.hpp
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef CENTERPOINT_TRT_HPP_
-#define CENTERPOINT_TRT_HPP_
+#ifndef LIDAR_CENTERPOINT__CENTERPOINT_TRT_HPP_
+#define LIDAR_CENTERPOINT__CENTERPOINT_TRT_HPP_
 
-#include <cuda_utils.hpp>
-#include <network_trt.hpp>
-#include <postprocess_kernel.hpp>
-#include <voxel_generator.hpp>
+#include <lidar_centerpoint/cuda_utils.hpp>
+#include <lidar_centerpoint/network/network_trt.hpp>
+#include <lidar_centerpoint/postprocess/postprocess_kernel.hpp>
+#include <lidar_centerpoint/preprocess/voxel_generator.hpp>
 
 #include <sensor_msgs/msg/point_cloud2.hpp>
 
@@ -107,4 +107,4 @@ private:
 
 }  // namespace centerpoint
 
-#endif  // CENTERPOINT_TRT_HPP_
+#endif  // LIDAR_CENTERPOINT__CENTERPOINT_TRT_HPP_
